+++ flutter_osm_interface/lib/src/osm_controller/osm_controller.dart
@@ -13,7 +13,9 @@ abstract class IBaseOSMController {
   /// [p] : (GeoPoint) position that will be added to map
   Future<void> changeLocation(GeoPoint p);
 
-  /// create marker int specific position without change map camera
+  /// addMarker
+  /// create marker int specific position without
+  /// change map camera
   ///
   /// [p] : (GeoPoint) desired location
   ///
@@ -24,14 +26,27 @@ abstract class IBaseOSMController {
     double? angle,
   });
 
-  ///remove marker from map of position
+  /// removeMarker
+  /// remove marker from map of position
+  ///
   /// [p] : geoPoint
   Future<void> removeMarker(GeoPoint p);
 
-  ///change Icon Marker
+
+  /// setIconMarker
+  /// this method change marker icon , marker should be already exist in the map
+  /// or it will throw exception that marker not exist
+  ///
+  /// [point]      : (geoPoint) geoPoint that want to change Icon
+  ///
+  /// [markerIcon] : (MarkerIcon) the new icon marker that will replace old icon
+  Future<void> setIconMarker(GeoPoint point,MarkerIcon markerIcon);
+
+
+  /// change Home Icon Marker
   /// we need to global key to recuperate widget from tree element
-  /// [key] : (GlobalKey) key of widget that represent the new marker
-  Future changeDefaultIconMarker(GlobalKey? key);
+  /// [homeMarker] : (MarkerIcon) key of widget that represent the new marker
+  Future changeDefaultIconMarker(MarkerIcon homeMarker);
 
   ///change  Marker of specific static points
   /// we need to global key to recuperate widget from tree element